fix schema definitions
