+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/schemas/SchemaDefinitions.java
@@ -74,9 +74,7 @@ public interface SchemaDefinitions {
 
   AggregateAndProofSchema getAggregateAndProofSchema();
 
-  default AttestationSchema<?> getAttestationSchema() {
-    return getAggregateAndProofSchema().getAttestationSchema();
-  }
+  AttestationSchema<?> getAttestationSchema();
 
   IndexedAttestationSchema getIndexedAttestationSchema();
