+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/schemas/SchemaDefinitionsBellatrix.java
@@ -54,9 +54,9 @@ public class SchemaDefinitionsBellatrix extends SchemaDefinitionsAltair {
 
   public SchemaDefinitionsBellatrix(final SpecConfigBellatrix specConfig) {
     super(specConfig);
+    final long maxValidatorsPerAttestation = getMaxValidatorPerAttestation(specConfig);
     this.beaconStateSchema = BeaconStateSchemaBellatrix.create(specConfig);
     this.executionPayloadHeaderSchema = beaconStateSchema.getLastExecutionPayloadHeaderSchema();
-    final long maxValidatorsPerAttestation = getMaxValidatorPerAttestation(specConfig);
     this.beaconBlockBodySchema =
         BeaconBlockBodySchemaBellatrixImpl.create(
             specConfig,
